+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 #Release Notes
 
 ## v0.2.0
+- Breaking change: fix typo: hiearchy to hierarchy #9
 - Added dependency on Powershell version 6.0 and higher.
 - Added support of Exporting to Json format.
 - Minor bugfixes.